+++ inst/doc/sps.R
@@ -1,9 +1,3 @@
-## ----include = FALSE----------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-
 ## ----frame--------------------------------------------------------------------
 library(sps)
 set.seed(123654)
@@ -15,12 +9,15 @@ frame <- data.frame(
 
 head(frame)
 
+
 ## ----outcome------------------------------------------------------------------
 sales <- round(frame$revenue * runif(1e3, 0.5, 2))
+
 
 ## ----allocation---------------------------------------------------------------
 allocation <- with(frame, prop_allocation(revenue, 100, region))
 allocation
+
 
 ## ----sample-------------------------------------------------------------------
 sample <- with(frame, sps(revenue, allocation, region))
@@ -29,17 +26,21 @@ survey <- cbind(frame[sample, ], sales = sales[sample])
 
 head(survey)
 
+
 ## ----weights------------------------------------------------------------------
 survey$weight <- weights(sample)
 
 head(survey)
 
+
 ## ----estimate-----------------------------------------------------------------
 ht <- with(survey, sum(sales * weight))
 ht
 
+
 ## ----bias---------------------------------------------------------------------
 ht / sum(sales) - 1
+
 
 ## ----variance-----------------------------------------------------------------
 repweights <- sps_repweights(weights(sample))
@@ -48,6 +49,7 @@ var <- attr(repweights, "tau")^2 *
   mean((colSums(survey$sales * repweights) - ht)^2)
 
 sqrt(var) / ht
+
 
 ## ----variance2----------------------------------------------------------------
 sps_var <- function(y, w) {
@@ -65,10 +67,12 @@ var <- with(
 
 sqrt(sum(var)) / ht
 
+
 ## ----prns---------------------------------------------------------------------
 frame$prn <- runif(1000)
 
 head(frame)
+
 
 ## ----prn samples--------------------------------------------------------------
 pareto <- order_sampling(\(x) x / (1 - x))
@@ -79,12 +83,14 @@ parsample <- with(frame, pareto(revenue, allocation, region, (prn - 0.5) %% 1))
 
 length(intersect(sample, parsample)) / 100
 
+
 ## ----prn simualtion-----------------------------------------------------------
 replicate(1000, {
   s <- with(frame, pareto(revenue, allocation, region))
   length(intersect(sample, s)) / 100
 }) |>
   summary()
+
 
 ## ----top up-------------------------------------------------------------------
 sample <- with(frame, sps(revenue, allocation, region, prn))
@@ -93,8 +99,10 @@ sample_tu <- with(frame, sps(revenue, allocation + c(10, 0, 0), region, prn))
 
 all(sample %in% sample_tu)
 
+
 ## ----critical-----------------------------------------------------------------
 Map(\(x) head(becomes_ta(x)), split(frame$revenue, frame$region))
+
 
 ## ----no_tu--------------------------------------------------------------------
 set.seed(13026)
@@ -107,8 +115,30 @@ sample <- sps(x, 4, prn = u)
 
 sample %in% sps(x, 5, prn = u)
 
+
 ## ----yes_tu-------------------------------------------------------------------
 sample %in% sps(x, 6, prn = u)
+
+
+## -----------------------------------------------------------------------------
+s <- sps_iterator(x, prn = u)
+for (i in 1:5) {
+  print(s())
+}
+
+sps(x, 6, prn = u)
+
+
+## -----------------------------------------------------------------------------
+set.seed(10052)
+u <- runif(10)
+pareto(x, 2, prn = u) %in% pareto(x, 3, prn = u)
+
+set.seed(10063)
+u <- runif(10)
+successive <- order_sampling(\(x) log(1 - x))
+successive(x, 2, prn = u) %in% successive(x, 3, prn = u)
+
 
 ## ----ht bias------------------------------------------------------------------
 sampling_distribution <- replicate(1000, {
@@ -118,7 +148,11 @@ sampling_distribution <- replicate(1000, {
 
 summary(sampling_distribution / sum(sales) - 1)
 
+
 ## ----tille, fig.width=8, fig.height=5.33--------------------------------------
+#| fig.alt: >
+#|   Empirical distribution of inclusion probabilities under sequential Poisson
+#|   sampling is approximately Guassian.
 set.seed(123456)
 n <- 5e3
 frame1 <- subset(frame, region == 1)
@@ -126,16 +160,18 @@ frame1 <- subset(frame, region == 1)
 pi_est <- tabulate(
   replicate(n, sps(frame1$revenue, allocation[1])),
   nbins = nrow(frame1)
-) / n
+)
 
 pi <- inclusion_prob(frame1$revenue, allocation[1])
 
-dist <- (pi_est - pi) / sqrt(pi * (1 - pi) / n)
+dist <- (pi_est / n - pi) / sqrt(pi * (1 - pi) / n)
 
 plot(
   density(dist, na.rm = TRUE),
-  ylim = c(0, 0.5), xlim = c(-4, 4),
-  ylab = "", xlab = "",
+  ylim = c(0, 0.5),
+  xlim = c(-4, 4),
+  ylab = "",
+  xlab = "",
   main = "Empirical distribution of inclusion probabilities"
 )
 lines(seq(-4, 4, 0.1), dnorm(seq(-4, 4, 0.1)), lty = "dashed")